requirements/dev.txt
-r base.txt
django-debug-toolbar==1.9.1
django-extensions==2.0.7